+++ tests/dummy/app/Http/Controllers/PostsController.php
@@ -113,4 +113,130 @@ class PostsController extends JsonApiController
         event(new ResourceEvent('deleted', $record->getKey(), $request));
     }
 
+    /**
+     * @param Post $record
+     * @param ValidatedRequest $request
+     */
+    public function readingRelationship(Post $record, ValidatedRequest $request)
+    {
+        event(new ResourceEvent('reading-relationship', $record, $request));
+    }
+
+    /**
+     * @param Post $record
+     * @param ValidatedRequest $request
+     */
+    public function readingAuthor(Post $record, ValidatedRequest $request)
+    {
+        event(new ResourceEvent('reading-author', $record, $request));
+    }
+
+    /**
+     * @param Post $record
+     * @param ValidatedRequest $request
+     */
+    public function replacing(Post $record, ValidatedRequest $request)
+    {
+        event(new ResourceEvent('replacing', $record, $request));
+    }
+
+    /**
+     * @param Post $record
+     * @param ValidatedRequest $request
+     */
+    public function replacingAuthor(Post $record, ValidatedRequest $request)
+    {
+        event(new ResourceEvent('replacing-author', $record, $request));
+    }
+
+    /**
+     * @param Post $record
+     * @param ValidatedRequest $request
+     */
+    public function replaced(Post $record, ValidatedRequest $request)
+    {
+        event(new ResourceEvent('replaced', $record, $request));
+    }
+
+    /**
+     * @param Post $record
+     * @param ValidatedRequest $request
+     */
+    public function replacedAuthor(Post $record, ValidatedRequest $request)
+    {
+        event(new ResourceEvent('replaced-author', $record, $request));
+    }
+
+    /**
+     * @param Post $record
+     * @param ValidatedRequest $request
+     */
+    public function adding(Post $record, ValidatedRequest $request)
+    {
+        event(new ResourceEvent('adding', $record, $request));
+    }
+
+    /**
+     * @param Post $record
+     * @param ValidatedRequest $request
+     */
+    public function addingTags(Post $record, ValidatedRequest $request)
+    {
+        event(new ResourceEvent('adding-tags', $record, $request));
+    }
+
+    /**
+     * @param Post $record
+     * @param ValidatedRequest $request
+     */
+    public function addedTags(Post $record, ValidatedRequest $request)
+    {
+        event(new ResourceEvent('added-tags', $record, $request));
+    }
+
+    /**
+     * @param Post $record
+     * @param ValidatedRequest $request
+     */
+    public function added(Post $record, ValidatedRequest $request)
+    {
+        event(new ResourceEvent('added', $record, $request));
+    }
+
+    /**
+     * @param Post $record
+     * @param ValidatedRequest $request
+     */
+    public function removing(Post $record, ValidatedRequest $request)
+    {
+        event(new ResourceEvent('removing', $record, $request));
+    }
+
+    /**
+     * @param Post $record
+     * @param ValidatedRequest $request
+     */
+    public function removingTags(Post $record, ValidatedRequest $request)
+    {
+        event(new ResourceEvent('removing-tags', $record, $request));
+    }
+
+    /**
+     * @param Post $record
+     * @param ValidatedRequest $request
+     */
+    public function removedTags(Post $record, ValidatedRequest $request)
+    {
+        event(new ResourceEvent('removed-tags', $record, $request));
+    }
+
+    /**
+     * @param Post $record
+     * @param ValidatedRequest $request
+     */
+    public function removed(Post $record, ValidatedRequest $request)
+    {
+        event(new ResourceEvent('removed', $record, $request));
+    }
+
 }